requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
DateTime==4.3
distlib==0.3.4
Django==3.2.10
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
numpy==1.21.5
pandas==1.1.5
pipreqs==0.4.11
platformdirs==2.4.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.0
virtualenvwrapper-win==1.2.7
yarg==0.1.9
zipp==3.7.0
zope.interface==5.4.0
django-ipware
line-bot-sdk
psycopg2
django_crontab
django-cors-headers
django-environ
python-crontab